+++ solverModules/GIMME_solver_modules/IntegerPartitionGraph.cpp
@@ -26,7 +26,7 @@ IntegerPartitionGraph::IntegerPartitionGraph(vector<vector<Subspace>> subspaces,
 						vector<int> sortedParts1 = nodes[level][i]->integerPartition.partsSortedAscendingly;
 						vector<int> sortedParts2 = nodes[level+1][i]->integerPartition.partsSortedAscendingly;
 						int partThatWasSplit = -1;
-						for (unsigned long long k = sortedParts1.size() - 1; k >= 0; k--)
+						for (int k = (int)sortedParts1.size() - 1; k >= 0; k--)
 							if (sortedParts1[k] != sortedParts2[k + 1]) {
 								partThatWasSplit = sortedParts1[k];
 								break;
@@ -46,7 +46,7 @@ IntegerPartitionGraph::IntegerPartitionGraph(vector<vector<Subspace>> subspaces,
 IntegerPartitionGraph::~IntegerPartitionGraph() 
 {
 	for (int i = 0; i < nodes.size(); i++) {
-		unsigned long long nodeSize = nodes[i].size();
+		int nodeSize = (int)nodes[i].size();
 		for (int j = 0; j < nodeSize; j++) {
 			if (i == 0 && j == 0)
 				continue;
@@ -60,10 +60,10 @@ vector<Node*> IntegerPartitionGraph::getReachableNodes(Node* node)
 {
 	if (node->edgesFromThisNode.empty()) return vector<Node*>();
 
-	unsigned long long numOfIntegersInNode = node->integerPartition.partsSortedAscendingly.size();
+	int numOfIntegersInNode = (int)node->integerPartition.partsSortedAscendingly.size();
 
 	node->tempIntegerRoots = vector<int>();
-	for (unsigned long long level = numOfIntegersInNode; level < nodes.size(); level++)
+	for (int level = numOfIntegersInNode; level < nodes.size(); level++)
 		for (int i = 0; i < nodes[level].size(); i++)
 		{
 			nodes[level][i]->tempFlag = false;
@@ -76,7 +76,7 @@ vector<Node*> IntegerPartitionGraph::getReachableNodes(Node* node)
 	}
 
 	int numOfReachableNodes = 0;
-	for (unsigned long long level = numOfIntegersInNode; level < nodes.size() - 1; level++)
+	for (int level = numOfIntegersInNode; level < nodes.size() - 1; level++)
 	{
 		for (int i = 0; i < nodes[level].size(); i++)
 		{
@@ -98,7 +98,7 @@ vector<Node*> IntegerPartitionGraph::getReachableNodes(Node* node)
 
 	int index = 0;
 	vector<Node*> listOfReachableNodes(numOfReachableNodes);
-	for (unsigned long long level = numOfIntegersInNode; level < nodes.size() - 1; level++)
+	for (int level = numOfIntegersInNode; level < nodes.size() - 1; level++)
 	{
 		for (int i = 0; i < nodes[level].size(); i++) 
 			if (nodes[level][i]->tempFlag)
@@ -128,7 +128,7 @@ void IntegerPartitionGraph::updateEdges(int numPlayers, int largestIntegerBeingS
 			if (!listOfDirectlyConnectedIntegerPartitions.empty())
 			{
 
-				unsigned long long index;
+				int index;
 				if (nodes[level][i]->edgesFromThisNode.empty())
 				{
 
@@ -138,7 +138,7 @@ void IntegerPartitionGraph::updateEdges(int numPlayers, int largestIntegerBeingS
 				else
 				{
 
-					index = nodes[level][i]->edgesFromThisNode.size();
+					index = (int)nodes[level][i]->edgesFromThisNode.size();
 					std::vector<Edge> tempListOfEdges(nodes[level][i]->edgesFromThisNode.size() + listOfDirectlyConnectedIntegerPartitions.size());
 					for (int j = 0; j < nodes[level][i]->edgesFromThisNode.size(); j++)
 						tempListOfEdges[j] = nodes[level][i]->edgesFromThisNode[j];
@@ -155,7 +155,7 @@ void IntegerPartitionGraph::updateEdges(int numPlayers, int largestIntegerBeingS
 						std::vector<int> sortedParts1 = nodes[level][i]->integerPartition.partsSortedAscendingly;
 						std::vector<int> sortedParts2 = nodes[level+1][j]->integerPartition.partsSortedAscendingly;
 						int partThatWasSplit = -1;
-						for (unsigned long long k = sortedParts1.size() - 1; k >= 0; k--)
+						for (int k = (int)sortedParts1.size() - 1; k >= 0; k--)
 						{
 							if (sortedParts1[k] != sortedParts2[k + 1])
 							{
